+++ cmd/cmd_linux.go
@@ -114,6 +114,8 @@ func coreSupportsReExec(corePath string) bool {
 	return true
 }
 
+// TODO: move to cmd/cmdutil/
+//
 // InternalToolPath returns the path of an internal snapd tool. The tool
 // *must* be located inside the same tree as the current binary.
 //
