+++ osutil/exec_test.go
@@ -23,11 +23,8 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"os/exec"
-	"path/filepath"
-	"strings"
 	"syscall"
 	"time"
 
@@ -48,69 +45,6 @@ func (s *execSuite) SetUpTest(c *C) {
 
 func (s *execSuite) TearDownTest(c *C) {
 	dirs.SetRootDir("")
-}
-
-func (s *execSuite) makeMockLdSoConf(c *C) {
-	ldSoConf := filepath.Join(dirs.SnapMountDir, "/core/current/etc/ld.so.conf")
-	ldSoConfD := ldSoConf + ".d"
-
-	err := os.MkdirAll(filepath.Dir(ldSoConf), 0755)
-	c.Assert(err, IsNil)
-	err = os.MkdirAll(ldSoConfD, 0755)
-	c.Assert(err, IsNil)
-
-	err = ioutil.WriteFile(ldSoConf, []byte("include /etc/ld.so.conf.d/*.conf"), 0644)
-	c.Assert(err, IsNil)
-
-	ldSoConf1 := filepath.Join(ldSoConfD, "x86_64-linux-gnu.conf")
-
-	err = ioutil.WriteFile(ldSoConf1, []byte(`
-# Multiarch support
-/lib/x86_64-linux-gnu
-/usr/lib/x86_64-linux-gnu`), 0644)
-	c.Assert(err, IsNil)
-}
-
-func (s *execSuite) TestCommandFromCore(c *C) {
-	s.makeMockLdSoConf(c)
-	root := filepath.Join(dirs.SnapMountDir, "/core/current")
-
-	os.MkdirAll(filepath.Join(root, "/usr/bin"), 0755)
-	osutil.CopyFile(truePath, filepath.Join(root, "/usr/bin/xdelta3"), 0)
-	cmd, err := osutil.CommandFromCore(dirs.SnapMountDir, "/usr/bin/xdelta3", "--some-xdelta-arg")
-	c.Assert(err, IsNil)
-
-	out, err := exec.Command("/bin/sh", "-c", fmt.Sprintf("readelf -l %s |grep interpreter:|cut -f2 -d:|cut -f1 -d]", truePath)).Output()
-	c.Assert(err, IsNil)
-	interp := strings.TrimSpace(string(out))
-
-	c.Check(cmd.Args, DeepEquals, []string{
-		filepath.Join(root, interp),
-		"--library-path",
-		fmt.Sprintf("%s/lib/x86_64-linux-gnu:%s/usr/lib/x86_64-linux-gnu", root, root),
-		filepath.Join(dirs.SnapMountDir, "/core/current/usr/bin/xdelta3"),
-		"--some-xdelta-arg",
-	})
-}
-
-func (s *execSuite) TestCommandFromCoreSymlinkCycle(c *C) {
-	s.makeMockLdSoConf(c)
-	root := filepath.Join(dirs.SnapMountDir, "/core/current")
-
-	os.MkdirAll(filepath.Join(root, "/usr/bin"), 0755)
-	osutil.CopyFile(truePath, filepath.Join(root, "/usr/bin/xdelta3"), 0)
-
-	out, err := exec.Command("/bin/sh", "-c", "readelf -l /bin/true |grep interpreter:|cut -f2 -d:|cut -f1 -d]").Output()
-	c.Assert(err, IsNil)
-	interp := strings.TrimSpace(string(out))
-
-	coreInterp := filepath.Join(root, interp)
-	c.Assert(os.MkdirAll(filepath.Dir(coreInterp), 0755), IsNil)
-	c.Assert(os.Symlink(filepath.Base(coreInterp), coreInterp), IsNil)
-
-	_, err = osutil.CommandFromCore(dirs.SnapMountDir, "/usr/bin/xdelta3", "--some-xdelta-arg")
-	c.Assert(err, ErrorMatches, "cannot run command from core: symlink cycle found")
-
 }
 
 func (s *execSuite) TestRunAndWaitRunsAndWaits(c *C) {
