+++ tests/lib/prepare.sh
@@ -307,6 +307,17 @@ prepare_classic() {
     fi
 }
 
+repack_snapd_snap_with_deb_content() {
+    local TARGET="$1"
+
+    local UNPACK_DIR="/tmp/snapd-unpack"
+    unsquashfs -no-progress -d "$UNPACK_DIR" snapd_*.snap
+    dpkg-deb -x "$SPREAD_PATH"/../snapd_*.deb "$UNPACK_DIR"
+    cp /usr/lib/snapd/info "$UNPACK_DIR"/usr/lib/
+    snap pack "$UNPACK_DIR" "$TARGET"
+    rm -rf "$UNPACK_DIR"
+}
+
 setup_reflash_magic() {
     # install the stuff we need
     distro_install_package kpartx busybox-static
@@ -340,12 +351,8 @@ setup_reflash_magic() {
     export UBUNTU_IMAGE_SNAP_CMD="$IMAGE_HOME/snap"
 
     if is_core18_system; then
-        # modify the snapd snap so that it has our snapd
-        UNPACK_DIR="/tmp/snapd-snap"
-        unsquashfs -no-progress -d "$UNPACK_DIR" snapd_*.snap
-        dpkg-deb -x "$SPREAD_PATH"/../snapd_*.deb "$UNPACK_DIR"
-        snap pack "$UNPACK_DIR" "$IMAGE_HOME"
-        
+        repack_snapd_snap_with_deb_content "$IMAGE_HOME"
+
         # FIXME: fetch directly once its in the assertion service
         cp "$TESTSLIB/assertions/ubuntu-core-18-amd64.model" "$IMAGE_HOME/pc.model"
         IMAGE=core18-amd64.img
