+++ tests/main/snapd-snap/task.yaml
@@ -7,7 +7,10 @@ systems: [ubuntu-16.04-64]
 priority: 100
 
 restore: |
+    echo "Cleanup the installed snapcraft"
     apt autoremove -y snapcraft
+    echo "Cleanup the build snapd snap"
+    rm -f "$PROJECT_PATH"/snapd_*.snap
 
 execute: |
     echo "Installing snapscraft"
@@ -19,3 +22,10 @@ execute: |
     echo "Ensure we have the fc-cache binaries"
     unsquashfs -ll snapd_*.snap | MATCH bin/fc-cache-v6
     unsquashfs -ll snapd_*.snap | MATCH bin/fc-cache-v7
+
+    echo "Ensure we have xdelta3"
+    unsquashfs -ll snapd_*.snap | MATCH bin/xdelta3
+
+    echo "Ensure we have mksquashfs (and the dependencies)"
+    unsquashfs -ll snapd_*.snap | MATCH bin/mksquashfs
+    unsquashfs -ll snapd_*.snap | MATCH liblzma.so.5
